move example scripts to examples
