+++ examples/settings.py
