--- modeling/settings.py
@@ -1,6 +1,0 @@
-import os
-import sys
-
-project_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-if project_root not in sys.path:
-    sys.path.append(project_root)
