+++ toy_project/dot_detector/check_swin_output.py
