+++ toy_project/dot_detector/dot_detector.py
