+++ toy_project/dot_detector/gen_dataset.py
