+++ toy_project/dot_detector/position_encoding.py
